+++ packages/components/table/src/index.vue
@@ -96,6 +96,7 @@
       scrollbar-always-on
       class="table"
       v-bind="$attrs"
+      :row-key="rowKey"
       @selection-change="handleSelectionChange"
       @sort-change="handleSortChange"
       @row-click="handleClickRow"
@@ -106,7 +107,8 @@
       <el-table-column v-if="isSelection" key="selection" type="selection" width="34" />
       <!-- 序号栏 -->
       <IndexColumn :show="isShowNumber" :sub-page-info="pagination.modelValue" align="left" />
-
+      <DragSortColumn :show-drag-sort="isShowDragSort" @subSortEnd="subSortEnd" />
+      <!-- <el-table-column v-if="isShowDragSort" label="排序">111</el-table-column> -->
       <!-- 展开行 -->
       <el-table-column v-if="hasExpand" type="expand">
         <template #default="{ row, $index }">
@@ -147,10 +149,12 @@ import PlusImagePreview from '@plus-pro-components/components/image-preview'
 import type { PlusPaginationProps } from '@plus-pro-components/components/pagination'
 import type { CSSProperties } from 'vue'
 import type { ComponentSize } from 'element-plus/es/constants'
-import PlusPopover from '../../popover/src/index.vue'
+import PlusPopover from '@plus-pro-components/components/popover'
+import type { PageInfo } from '@plus-pro-components/types'
 import PlusTableActionBar from './table-action-bar.vue'
 import CustomColumn from './table-column.vue'
 import IndexColumn from './table-column-index.vue'
+import DragSortColumn from './table-column-drag-sort.vue'
 import type {
   ButtonsCallBackParams,
   TableState,
@@ -189,22 +193,26 @@ export interface PlusTableProps {
   config: TableConfigRow[]
   /* 表格头样式*/
   headerCellStyle?: CSSProperties
+  // 是否可拖拽
+  isShowDragSort?: boolean
+  rowKey?: string
 }
 
 export interface PlusTableEmits {
   (e: 'subPaginationChange', pageInfo: PageInfo): void
-  (e: 'subSelected', data: any): void
-  (e: 'subCurrent', data: any): void
-  (e: 'subExpandChange', data: any): void
-  (e: 'subSortChange', data: any): void
-  (e: 'subClickRow', row: any, column: any, event: any): void
-  (e: 'subClickButton', data: any): void
-  (e: 'subClickButton'): void
+  (e: 'subSelected', data: any[]): void
+  (e: 'subCurrent', row: any): void
+  (e: 'subExpandChange', row: any): void
+  (e: 'subSortChange', sortParams: SortParams): void
+  (e: 'subClickRow', row: any, column: any, event: MouseEvent): void
+  (e: 'subClickButton', data: ButtonsCallBackParams): void
+  (e: 'subSortEnd', newIndex: number, oldIndex: number): void
 }
 
 defineOptions({
   name: 'PlusTable'
 })
+
 const LabelLength = 6
 const props = withDefaults(defineProps<PlusTableProps>(), {
   size: 'small',
@@ -234,7 +242,9 @@ const props = withDefaults(defineProps<PlusTableProps>(), {
   headerCellStyle: () => ({
     backgroundColor: '#F5F9FD',
     color: '#777'
-  })
+  }),
+  isShowDragSort: false,
+  rowKey: 'id'
 })
 
 const emit = defineEmits<PlusTableEmits>()
@@ -333,7 +343,7 @@ const handleSortChange = (sortParams: SortParams) => {
   emit('subSortChange', sortParams)
 }
 // 当某一行被点击时会触发该事件
-const handleClickRow = (row: any, column: any, event: PointerEvent) => {
+const handleClickRow = (row: any, column: any, event: MouseEvent) => {
   emit('subClickRow', row, column, event)
 }
 // 点击按钮传递给父组件
@@ -379,7 +389,7 @@ const getLabel = (label: string) => {
   return label?.slice(0, LabelLength) + '...'
 }
 // 多选处理
-const handleSelectionChange = (data: any) => {
+const handleSelectionChange = (data: any[]) => {
   emit('subSelected', data)
 }
 // 密度
@@ -392,6 +402,9 @@ const handleClickDensity = (data: string) => {
     }
   })
   state.size = data
+}
+const subSortEnd = (newIndex: number, oldIndex: number) => {
+  emit('subSortEnd', newIndex, oldIndex)
 }
 
 handleClickDensity(state.size || props.size || 'small')
@@ -407,7 +420,7 @@ const {
   isIndeterminate
 } = toRefs(state)
 
-// 暴露画布生成图片方法到外部调用
+// 暴露方法到外部调用
 defineExpose({
   scrollTo,
   doLayout,
@@ -449,6 +462,27 @@ defineExpose({
   .plus-table-title-expand-col {
     padding: 0;
   }
+
+  .el-table .cell {
+    line-height: 12px;
+  }
+  .el-table__body-wrapper .el-table__body .el-table__row td {
+    border-right: 0;
+  }
+  .el-table--border th.el-table__cell {
+    border: none;
+    ::before {
+      position: absolute;
+      top: 50%;
+      inset-inline-end: 0;
+      width: 1px;
+      height: 1.6em;
+      background-color: #ccc;
+      transform: translateY(-50%);
+      transition: background-color 0.2s;
+      content: '';
+    }
+  }
 }
 
 .plus-table-filter-dialog {
@@ -466,26 +500,6 @@ defineExpose({
     .el-checkbox:nth-of-type(4n) {
       margin: 0;
     }
-  }
-}
-.el-table .cell {
-  line-height: 12px;
-}
-.el-table .el-table__body-wrapper .el-table__body .el-table__row td {
-  border-right: 0 !important;
-}
-.el-table--border th.el-table__cell {
-  border: none;
-  ::before {
-    position: absolute;
-    top: 50%;
-    inset-inline-end: 0;
-    width: 1px;
-    height: 1.6em;
-    background-color: #ccc;
-    transform: translateY(-50%);
-    transition: background-color 0.2s;
-    content: '';
   }
 }
 </style>