+++ packages/constants/page.ts
@@ -1,3 +1,5 @@
+import type { PageInfo } from '@plus-pro-components/types'
+
 export const defaultPageSizeList = [10, 20, 30, 40, 50, 100, 200, 300, 400, 500]
 
 export const defaultPageInfo: PageInfo = {