+++ packages/play/src/views/field-item.vue
@@ -1,0 +1,1 @@
+<template>field-item</template>