+++ packages/plus-pro-components/index.ts
@@ -1,4 +1,5 @@
 import installer from './defaults'
+
 export * from '@plus-pro-components/components'
 export * from '@plus-pro-components/hooks'