+++ packages/types/plus.ts
@@ -1,0 +1,28 @@
+export {}
+
+/**
+ * 分页参数
+ */
+export interface PageInfo {
+  /**
+   * 默认为1
+   */
+  page: number
+  /**
+   * 默认为10
+   */
+  pageSize: number
+  /**
+   * 总数
+   */
+  total?: number
+}
+
+/**
+ * 选择框类型
+ */
+export interface OptionsRow {
+  label: number | string
+  value: number | string
+  color?: string
+}